PasswordGenerator
I created a password generator that has 6 options that can be marked such as: lowercase letters, uppercase letters, numbers, symbols, spaces, and whether or not to duplicate characters.
In addition, it is possible to choose the length of the required password and receive feedback on the strength of the password according to its length.